+++ dst/app/routes/transactions/placeOrder.js
@@ -61,7 +61,7 @@ placeOrderTransactionsRouter.post('/start', permitScopes_1.default(['transaction
 (req, _, next) => {
     if (typeof req.body.sellerId === 'string') {
         req.body.seller = {
-            typeOf: cinerino.factory.organizationType.MovieTheater,
+            typeOf: cinerino.factory.chevre.organizationType.MovieTheater,
             id: req.body.sellerId
         };
     }