+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - ムビチケ認証処理をChevreへ移行
+- 旧メンバーシップ注文タスク(registerProgramMembership)を停止
 
 ### Deprecated