+++ CHANGELOG.md
@@ -18,6 +18,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v5.29.0 - 2020-09-10
+
+### Changed
+
+- ムビチケ認証処理をChevreへ移行
+- 旧メンバーシップ注文タスク(registerProgramMembership)を停止
+- @chevre/factoryと重複するインターフェースを最適化
+
 ## v5.28.0 - 2020-09-08
 
 ### Added